+++ tests/tests/src/ExternalArity.mjs
@@ -37,10 +37,13 @@ let ReactTest = {
   FormattedMessage: FormattedMessage
 };
 
+let ExternalArity$ReactTest$FormattedMessage$make = ReactIntl.FormattedMessage;
+
 export {
   f1,
   f2,
   FromTypeConstructor,
   ReactTest,
+  ExternalArity$ReactTest$FormattedMessage$make,
 }
 /* test1 Not a pure module */