+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 #### :bug: Bug fix
 
+- JSX v4 nested React component tags now use generated direct component exports when available, while preserving normal nested module exports. https://github.com/rescript-lang/rescript/pull/8293
 - Fix directive `@warning("-102")` not working. https://github.com/rescript-lang/rescript/pull/8322
 - Fix duplicated comments in `for`..`of` formatter. https://github.com/rescript-lang/rescript/pull/8395
 - Fix issue where warning 56 would blow up with `dict{}` patterns. https://github.com/rescript-lang/rescript/pull/8403